:truck: move SharedProfileForCampaign model to prescription context
